Remove the circular dependency between widget base and vdom
