+++ tests/widget-core/unit/vdom.ts
@@ -5,10 +5,10 @@ import { match, spy, stub, SinonSpy } from 'sinon';
 import { createResolvers } from './../support/util';
 import sendEvent from '../support/sendEvent';
 
-import { renderer, widgetInstanceMap } from '../../../src/widget-core/vdom';
+import { renderer } from '../../../src/widget-core/vdom';
 import { v, w, dom as d, VNODE } from '../../../src/widget-core/d';
 import { VNode, DNode, DomVNode } from '../../../src/widget-core/interfaces';
-import { WidgetBase } from '../../../src/widget-core/WidgetBase';
+import { WidgetBase, widgetInstanceMap } from '../../../src/widget-core/WidgetBase';
 import Registry from '../../../src/widget-core/Registry';
 import { I18nMixin } from '../../../src/widget-core/mixins/I18n';
 import registry from '../../../src/widget-core/decorators/registry';
